move theme update logic to SettingsVM
